changed type to custom on used by rp2040 sdk

+++ ncp3901.h
@@ -3,7 +3,7 @@
 
 void ncp3901_init(uint gpio_wireless_available, uint gpio_otg);
 
-void on_wireless_enabled(uint gpio, ulong events);
+void on_wireless_enabled(uint gpio, uint32_t events);
 
 void ncp3901_adc0();
